refactor(core): extract mutation request guard module
