+++ packages/core/src/vite/comments-plugin.ts
@@ -6,6 +6,7 @@ import { parse as babelParse } from '@babel/parser';
 import * as t from '@babel/types';
 import type { Connect, Plugin, ViteDevServer } from 'vite';
 import { walkAll, walkJsx } from './babel-walk.ts';
+import { validateMutationRequest } from './request-guard.ts';
 
 const MARKER_RE =
   /\{\/\*\s*@slide-comment\s+id="(c-[a-f0-9]+)"\s+ts="([^"]+)"\s+text="([A-Za-z0-9_-]+={0,2})"\s*\*\/\}/g;
@@ -65,87 +66,6 @@ function json(res: ServerResponse, status: number, body: unknown) {
   res.statusCode = status;
   res.setHeader('content-type', 'application/json');
   res.end(JSON.stringify(body));
-}
-
-type MutationRequestValidationResult =
-  | { ok: true }
-  | { ok: false; status: number; error: string };
-
-function firstHeaderValue(value: string | string[] | undefined): string | null {
-  if (Array.isArray(value)) return value[0] ?? null;
-  return value ?? null;
-}
-
-function headerValue(req: Connect.IncomingMessage, name: string): string | null {
-  return firstHeaderValue(req.headers[name.toLowerCase()])?.trim() ?? null;
-}
-
-function firstCommaToken(value: string | null): string | null {
-  if (!value) return null;
-  const [first] = value.split(',', 1);
-  return first?.trim() || null;
-}
-
-function requestProto(req: Connect.IncomingMessage): 'http' | 'https' {
-  const forwarded = firstCommaToken(headerValue(req, 'x-forwarded-proto'))?.toLowerCase();
-  if (forwarded === 'http' || forwarded === 'https') return forwarded;
-  return req.socket.encrypted ? 'https' : 'http';
-}
-
-function normalizedOrigin(origin: string): string | null {
-  try {
-    const url = new URL(origin);
-    return `${url.protocol}//${url.host}`.toLowerCase();
-  } catch {
-    return null;
-  }
-}
-
-export function validateMutationRequest(
-  req: Connect.IncomingMessage,
-  opts: { requireJsonBody?: boolean } = {},
-): MutationRequestValidationResult {
-  if (opts.requireJsonBody) {
-    const contentType = headerValue(req, 'content-type')?.toLowerCase();
-    if (!contentType || !contentType.startsWith('application/json')) {
-      return {
-        ok: false,
-        status: 415,
-        error: 'content-type must be application/json',
-      };
-    }
-  }
-
-  const fetchSite = firstCommaToken(headerValue(req, 'sec-fetch-site'))?.toLowerCase();
-  if (fetchSite === 'cross-site') {
-    return { ok: false, status: 403, error: 'cross-site request blocked' };
-  }
-
-  const originRaw = headerValue(req, 'origin');
-  if (!originRaw) return { ok: true };
-  if (originRaw.toLowerCase() === 'null') {
-    return { ok: false, status: 403, error: 'opaque origin is not allowed' };
-  }
-
-  const actualOrigin = normalizedOrigin(originRaw);
-  if (!actualOrigin) {
-    return { ok: false, status: 403, error: 'invalid origin header' };
-  }
-
-  const host = firstCommaToken(headerValue(req, 'x-forwarded-host')) ?? headerValue(req, 'host');
-  if (!host) {
-    return { ok: false, status: 400, error: 'missing host header' };
-  }
-  const expectedOrigin = `${requestProto(req)}://${host}`.toLowerCase();
-  if (actualOrigin !== expectedOrigin) {
-    return {
-      ok: false,
-      status: 403,
-      error: 'origin mismatch',
-    };
-  }
-
-  return { ok: true };
 }
 
 function resolveSlidePath(userCwd: string, slidesDir: string, slideId: string): string | null {
